Add more compiler warning flags
Enable more compiler flags suggested from:

https://github.com/lefticus/cppbestpractices

+++ include/string_format.tcc
@@ -8,9 +8,9 @@ namespace buildsys {
 	 */
 	template < typename ... Args >
 	    std::string string_format(const std::string & format, Args ... args) {
-		size_t size = snprintf(nullptr, 0, format.c_str(), args ...) + 1;	// Extra space for '\0'
+		int size = snprintf(nullptr, 0, format.c_str(), args ...) + 1;	// Extra space for '\0'
 		std::unique_ptr < char[] > buf(new char[size]);
-		snprintf(buf.get(), size, format.c_str(), args ...);
+		snprintf(buf.get(), static_cast<size_t>(size), format.c_str(), args ...);
 		return std::string(buf.get(), buf.get() + size - 1);	// We don't want the '\0' inside
 	}
 }
